+++ CHANGES.md
@@ -11,6 +11,7 @@
 * refactor internals for improved readability
 * show classes that contain no lines, e.g. `__init__.py`
 * add `Cobertura.filename(class_name)` to retrieve the filename of a class
+* show source code in HTML report
 
 ## 0.2.1 (2014-12-10)